Switch from own geometry types to kurbo 🎋

+++ scale.rs
@@ -9,20 +9,17 @@ pub enum Scale<T> {
     /// An absolute value.
     Abs(T),
     /// A relative value.
-    Rel(f32),
+    Rel(f64),
 }
 
-impl<T> Scale<T> where T: Mul<f32, Output=T> {
+impl<T> Scale<T> where T: Mul<f64, Output=T> {
     /// Returns either the absolute value or computes the relative value as a
     /// fraction of `one`.
     ///
     /// # Example
     /// ```
-    /// # use layr::{assert_approx_eq, geom::{Length, Scale}};
-    /// assert_approx_eq!(
-    ///     Scale::Rel(0.5).resolve(Length::cm(5.0)),
-    ///     Length::cm(2.5),
-    /// );
+    /// # use layr::{assert_approx_eq, geom::Scale};
+    /// assert_approx_eq!(Scale::Rel(0.5).resolve(5.0), 2.5);
     /// ```
     pub fn resolve(self, one: T) -> T {
         match self {
@@ -33,7 +30,7 @@ impl<T> Scale<T> where T: Mul<f32, Output=T> {
 }
 
 impl<T> ApproxEq for Scale<T> where T: ApproxEq {
-    fn approx_eq(&self, other: &Self, tolerance: f32) -> bool {
+    fn approx_eq(&self, other: &Self, tolerance: f64) -> bool {
         match (self, other) {
             (Scale::Abs(x), Scale::Abs(y)) => x.approx_eq(y, tolerance),
             (Scale::Rel(x), Scale::Rel(y)) => x.approx_eq(y, tolerance),
